requirements.txt
Django>=3.1.3
google-search-results-serpwow>=1.1.11